requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0
albumentations>=1.3.0

# API & Deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Docker & Production
gunicorn>=21.2.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
